Convert runner/clireporter.js to typescript
